refactor - signup added to routes
