+++ src/components/Signup/index.js
@@ -1,10 +1,8 @@
-import Signup from "./Signup";
 import SignupHeader from "./SignupHeader";
 import SignupCard from "./SignupCard";
 import SignupForm from "./SignupForm";
 
 export {
-    Signup,
     SignupHeader,
     SignupCard,
     SignupForm
