+++ src/pages/index.js
@@ -2,10 +2,12 @@ import Cart from "./Cart";
 import Home from "./Home";
 import Wishlist from "./Wishlist";
 import Products from "./Products";
+import Signup from "./Signup";
 
 export {
     Cart,
     Home,
     Wishlist,
-    Products
+    Products,
+    Signup
 }
