+++ src/utils/index.js
@@ -1,7 +1,7 @@
-import { getTextColor, getBgColor, getSolidBtnBgColor, getSolidBtnTextColor, getBorderColor, getIconColor } from "./theme.util";
+import { getTextColor, getBgColor, getSolidBtnBgColor, getSolidBtnTextColor, getBorderColor, getIconColor, getBadgeBgColor, getBadgeTextColor } from "./theme.util";
 import { emailIsInvalid, nameIsInvalid, passIsInvalid, passAndConfPassAreDiff } from "./inputValidation.util";
 import { timer } from "./timer.util";
-import { getTotalDiscount, getTotalPrice, getTotalItems, getCartTotal, getTotalOfferPrice } from './cart.util'
+import { getTotalDiscount, getTotalPrice, getTotalCartItems, getCartTotal, getTotalOfferPrice } from './cart.util'
 import { getTotalWishlistItems } from './wishlist.util'
 import { isInputIncluded, isSortInputIncluded } from './filterChecks.util'
 import { filterProductsByBrand, filterProductsByGender, filterProductsByPrice, filterProductsByRating, filterProductsBySize } from './productsContext.util'
@@ -14,6 +14,8 @@ export {
     getSolidBtnBgColor,
     getSolidBtnTextColor,
     getIconColor,
+    getBadgeBgColor,
+    getBadgeTextColor,
     emailIsInvalid,
     nameIsInvalid,
     passAndConfPassAreDiff,
@@ -21,7 +23,7 @@ export {
     timer,
     getTotalDiscount,
     getCartTotal,
-    getTotalItems,
+    getTotalCartItems,
     getTotalOfferPrice,
     getTotalPrice,
     getTotalWishlistItems,
